レビュー訴求系の処理を共通化
