+++ app/src/main/java/wacode/yamada/yuki/nempaymentapp/view/activity/MainActivity.kt
@@ -37,7 +37,6 @@ import wacode.yamada.yuki.nempaymentapp.view.activity.drawer.MosaicListActivity
 import wacode.yamada.yuki.nempaymentapp.view.activity.drawer.RaccoonDonateActivity
 import wacode.yamada.yuki.nempaymentapp.view.adapter.ExampleFragmentPagerAdapter
 import wacode.yamada.yuki.nempaymentapp.view.controller.DrawerListController
-import wacode.yamada.yuki.nempaymentapp.view.dialog.RaccoonConfirmDialog
 import wacode.yamada.yuki.nempaymentapp.view.dialog.RaccoonConfirmViewModel
 import wacode.yamada.yuki.nempaymentapp.view.fragment.SplashFragment
 import wacode.yamada.yuki.nempaymentapp.view.fragment.top.SendTopFragment
@@ -219,45 +218,11 @@ class MainActivity : BaseActivity(), SplashCallback, QrScanCallback, DrawerListC
             setupViewPager()
             setupBottomTabLayout()
             setupNemIcon()
-            if (!isAlreadyShownReviewDialog() && RegisterDateUtils.isThreeDaysLater(this)) {
+            if (!ReviewAppealUtils.isAlreadyShownReviewDialog(this) && RegisterDateUtils.isThreeDaysLater(this)) {
                 val viewModel = RaccoonConfirmViewModel()
-                saveAlreadyShownReviewDialog()
-                createReviewDialog(viewModel)
+                ReviewAppealUtils.saveAlreadyShownReviewDialog(this)
+                ReviewAppealUtils.createReviewDialog(this, supportFragmentManager, viewModel)
             }
-        }
-    }
-
-    private fun createReviewDialog(viewModel: RaccoonConfirmViewModel) {
-        val title = getString(R.string.about_activity_review_title)
-        val message = getString(R.string.about_activity_review_message)
-        val buttonText = getString(R.string.about_activity_review_button)
-        viewModel.checkEvent
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe { item ->
-                    viewModel.saveSPTwiceDisplay(this, RaccoonConfirmViewModel.KEY_REVIEW, item)
-                }
-        viewModel.clickEvent
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe {
-                    openPlayStore()
-                }
-
-        val dialog = RaccoonConfirmDialog.createDialog(
-                viewModel,
-                title,
-                message,
-                buttonText,
-                viewModel.shouldTwiceDisplay(this, RaccoonConfirmViewModel.KEY_REVIEW)
-        )
-
-        dialog.show(supportFragmentManager, RaccoonConfirmDialog::class.java.toString())
-    }
-
-    private fun openPlayStore() {
-        try {
-            this.startActivity(Intent(Intent.ACTION_VIEW, Uri.parse("market://details?id=" + this.packageName)))
-        } catch (exception: android.content.ActivityNotFoundException) {
-            this.startActivity(Intent(Intent.ACTION_VIEW, Uri.parse("https://play.google.com/store/apps/details?id=" + this.packageName)))
         }
     }
 
@@ -344,15 +309,10 @@ class MainActivity : BaseActivity(), SplashCallback, QrScanCallback, DrawerListC
 
     private fun isAlreadyRaccoonWallet() = SharedPreferenceUtils[this, SP_IS_FIRST_RACCOON, true]
 
-    private fun isAlreadyShownReviewDialog() = SharedPreferenceUtils[this, SP_IS_ALREADY_SHOWN_REVIEW_DIALOG, false]
-
-    private fun saveAlreadyShownReviewDialog() = SharedPreferenceUtils[this, SP_IS_ALREADY_SHOWN_REVIEW_DIALOG, true]
-
     private fun saveRegisterDate() = RegisterDateUtils.saveRegisterDate(this)
 
     companion object {
         const val SP_IS_FIRST_RACCOON = "sp_is_first_raccoon"
-        private const val SP_IS_ALREADY_SHOWN_REVIEW_DIALOG = "sp_is_already_shown_review_dialog"
         private const val HOME_POSITION = 2
         private const val ARG_SHOULD_SHOW_SPLASH = "args_show_splash"
         fun createIntent(context: Context) = Intent(context, MainActivity::class.java)
