+++ app/src/main/java/wacode/yamada/yuki/nempaymentapp/view/activity/drawer/AboutActivity.kt
@@ -6,11 +6,10 @@ import android.net.Uri
 import android.os.Bundle
 import android.support.customtabs.CustomTabsIntent
 import com.google.android.gms.oss.licenses.OssLicensesMenuActivity
-import io.reactivex.android.schedulers.AndroidSchedulers
 import kotlinx.android.synthetic.main.activity_about.*
 import wacode.yamada.yuki.nempaymentapp.R
+import wacode.yamada.yuki.nempaymentapp.utils.ReviewAppealUtils
 import wacode.yamada.yuki.nempaymentapp.view.activity.BaseActivity
-import wacode.yamada.yuki.nempaymentapp.view.dialog.RaccoonConfirmDialog
 import wacode.yamada.yuki.nempaymentapp.view.dialog.RaccoonConfirmViewModel
 
 
@@ -67,45 +66,12 @@ class AboutActivity : BaseActivity() {
     private fun setupReviewButton() {
         plzReviewLayout.setOnClickListener {
             val viewModel = RaccoonConfirmViewModel()
-            if (viewModel.shouldTwiceDisplay(this, RaccoonConfirmViewModel.KEY_REVIEW)) {
-                createReviewDialog(viewModel)
+            if (viewModel.shouldTwiceDisplay(this, ReviewAppealUtils.KEY_REVIEW)) {
+                ReviewAppealUtils.saveAlreadyShownReviewDialog(this)
+                ReviewAppealUtils.createReviewDialog(this, supportFragmentManager, viewModel)
             } else {
-                openPlayStore()
+                ReviewAppealUtils.openPlayStore(this)
             }
-        }
-    }
-
-    private fun createReviewDialog(viewModel: RaccoonConfirmViewModel) {
-        val title = getString(R.string.about_activity_review_title)
-        val message = getString(R.string.about_activity_review_message)
-        val buttonText = getString(R.string.about_activity_review_button)
-        viewModel.checkEvent
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe { item ->
-                    viewModel.saveSPTwiceDisplay(this, RaccoonConfirmViewModel.KEY_REVIEW, !item)
-                }
-        viewModel.clickEvent
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe {
-                    openPlayStore()
-                }
-
-        val dialog = RaccoonConfirmDialog.createDialog(
-                viewModel,
-                title,
-                message,
-                buttonText,
-                viewModel.shouldTwiceDisplay(this, RaccoonConfirmViewModel.KEY_REVIEW)
-        )
-
-        dialog.show(supportFragmentManager, RaccoonConfirmDialog::class.java.toString())
-    }
-
-    private fun openPlayStore() {
-        try {
-            this.startActivity(Intent(Intent.ACTION_VIEW, Uri.parse("market://details?id=" + this.packageName)))
-        } catch (exception: android.content.ActivityNotFoundException) {
-            this.startActivity(Intent(Intent.ACTION_VIEW, Uri.parse("https://play.google.com/store/apps/details?id=" + this.packageName)))
         }
     }
